+++ test/nadia/config_test.exs
@@ -15,16 +15,12 @@ defmodule Nadia.ConfigTest do
     graph_base_url = Application.get_env(:nadia, :graph_base_url)
     file_base_url = Application.get_env(:nadia, :file_base_url)
     proxy = Application.get_env(:nadia, :proxy)
-    socks5_user = Application.get_env(:nadia, :socks5_user)
-    socks5_pass = Application.get_env(:nadia, :socks5_pass)
 
     on_exit(fn ->
       restore_env!(:base_url, base_url)
       restore_env!(:graph_base_url, graph_base_url)
       restore_env!(:file_base_url, file_base_url)
       restore_env!(:proxy, proxy)
-      restore_env!(:socks5_user, socks5_user)
-      restore_env!(:socks5_pass, socks5_pass)
     end)
   end
 
@@ -35,21 +31,9 @@ defmodule Nadia.ConfigTest do
   end
 
   test "Config.proxy/0 returns config value when present" do
-    :ok = Application.put_env(:nadia, :proxy, {:socks5, ~c"localhost", 1080})
+    :ok = Application.put_env(:nadia, :proxy, {:http, "localhost", 8080})
 
-    assert Config.proxy() == {:socks5, ~c"localhost", 1080}
-  end
-
-  test "Config.socks5_user/0 returns config value when present" do
-    :ok = Application.put_env(:nadia, :socks5_user, "user")
-
-    assert Config.socks5_user() == "user"
-  end
-
-  test "Config.socks5_pass/0 returns config value when present" do
-    :ok = Application.put_env(:nadia, :socks5_pass, "password")
-
-    assert Config.socks5_pass() == "password"
+    assert Config.proxy() == {:http, "localhost", 8080}
   end
 
   test "Config.base_url/0 returns environment variable" do